requirements.txt
streamlit
pillow
numpy
scikit-learn
matplotlib
pandas
torch
opencv-python-headless
scikit-learn
matplotlib
transformers